Click on LinkView selects it
